Code cleanup, submodule removed
