--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "webexteamssdk"]
-	path = webexteamssdk
-	url = https://github.com/CiscoDevNet/webexteamssdk.git
