--- config/ngrok.yaml
@@ -1,2 +1,0 @@
----
-web_addr: 0.0.0.0:4040
