+++ requirements.txt
@@ -4,4 +4,4 @@ pyopenssl
 ndg-httpsclient
 pyasn1
 pyaml
-./webexteamssdk
+webexteamssdk
